+++ FMM/FMM.csproj
@@ -1,8 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
-
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
-
-
+  <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.4"/>
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="3.1.4"/>
+    <PackageReference Include="Npgsql" Version="4.1.3.1"/>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.1.4"/>
+  </ItemGroup>
 </Project>
