--- FMM/Features/Dream.cs
@@ -1,8 +1,0 @@
-using System;
-
-public class Dream
-{
-    public Guid Id { get; set; }
-    public string Title { get; set; }
-    public string Description { get; set; }
-}
