+++ FMM/Persistent/Dream.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace FMM.Persistent
+{
+    public class Dream
+    {
+        public Guid Id { get; set; }
+        public string Title { get; set; }
+        public string Description { get; set; }
+    }
+}
